RangeShader + PlayerBase + Turret selection
Created a shader in shadergraph that will be applied to a sphere in order to generate the effect of the range.
Placed turret base on start game.
You can now select and deselect turrets properly. It should work for all of them and in the future add other functionality to it (like the upgrade system.